Move reservation to separate package Reduce tests spring context coupling
