+++ server/src/main/kotlin/org/itmodreamteam/myrest/server/service/reservation/ReservationService.kt
@@ -1,4 +1,4 @@
-package org.itmodreamteam.myrest.server.service.restaurant
+package org.itmodreamteam.myrest.server.service.reservation
 
 import org.itmodreamteam.myrest.server.model.restaurant.Reservation
 import org.itmodreamteam.myrest.server.model.restaurant.RestaurantTable
