+++ server/src/test/kotlin/org/itmodreamteam/myrest/server/service/restaurant/GetByIdTest.kt
@@ -31,9 +31,6 @@ class GetByIdTest {
     @MockBean
     lateinit var notificationService: NotificationService
 
-    @MockBean
-    lateinit var reservationService: ReservationService
-
     private var identifier: Long = 0
 
     @Before
