+++ server/src/test/kotlin/org/itmodreamteam/myrest/server/service/restaurant/UpdateStatusTest.kt
@@ -32,9 +32,6 @@ class UpdateStatusTest {
     @MockBean
     lateinit var notificationService: NotificationService
 
-    @MockBean
-    lateinit var reservationService: ReservationService
-
     private var identifier: Long = 0
 
     @Before
