+++ server/src/test/kotlin/org/itmodreamteam/myrest/server/service/restaurant/UpdateTest.kt
@@ -32,9 +32,6 @@ class UpdateTest {
     @MockBean
     lateinit var notificationService: NotificationService
 
-    @MockBean
-    lateinit var reservationService: ReservationService
-
     private lateinit var restaurant: Restaurant
     private var identifier: Long = 0
